+++ src/routes/(app)/workbench/panes/Inspector.svelte
